Modularize test cases into separate files
Make it easier to read, extend, and run separately.
