+++ tests/test_xml.py
@@ -1,0 +1,93 @@
+#
+
+from lxml import etree
+import pytest
+import re
+import os
+import sdsc
+
+
+def test_xml(xmltestcase):
+    """Runs specified testcase and validates the output of all checks.
+    The xmltestcase fixture returns all files in tests/cases"""
+    nr_errors = 0
+
+    testname = os.path.basename(xmltestcase)
+    try:
+        resultxml = sdsc.checkOneFile(xmltestcase)
+    except etree.Error as error:
+        pytest.fail("Syntax error in testcase {0!r}: {1}!".format(testname, error.msg))
+
+    # Parse the input file and gather all ids
+    inputtree = etree.parse(xmltestcase)
+    inputids = []
+    for elem in inputtree.getiterator():
+        eid = elem.get("{http://www.w3.org/XML/1998/namespace}id", elem.get("id"))
+        if eid is not None:
+            if eid.count("sdsc."):
+                if eid in inputids:
+                    pytest.fail(
+                        "Duplicate ID {0!r} in case {1!r}!".format(eid, testname))
+
+                inputids.append(eid)
+    if len(inputids) == 0:
+        pytest.skip("No tests found in {0}".format(testname))
+
+    # Parse the result file and collect ids of errors and warnings
+    resulttree = etree.fromstring(resultxml)
+    complaints = {}
+    currentPartSource = ""
+    for elem in resulttree.getiterator():
+        if elem.tag == "part":
+            currentPartSource = elem.get("source")
+            complaints[currentPartSource] = []
+        elif elem.tag == "result":
+            elemType = elem.get("type", "info")
+            if elemType == "info":
+                # Not interested in those. They don't have an ID either...
+                continue
+
+            withinid = elem.findtext("location/withinid")
+            message = elem.find("message")
+            if withinid is None:
+                withinid = elem.findtext("message/id")
+                if withinid is None:
+                    pytest.fail("No withinid found")
+
+            formattedMessage = "<no message>"
+            if message is not None:
+                formattedMessage = etree.tostring(
+                    message, method="text", encoding='UTF-8').decode(encoding='UTF-8')
+                # Remove excessive whitespace and newlines
+                formattedMessage = " ".join(formattedMessage.split()).strip()
+
+            complaints[currentPartSource].append(
+                {'id': withinid, 'message': formattedMessage, 'type': elemType})
+
+    # Isolate unexpected warnings
+    for checkmodule, complaintList in complaints.items():
+        for complaint in complaintList:
+            if not complaint["id"].count("sdsc.expect.{0}.{1}".format(complaint["type"], checkmodule)):
+                print("Unexpected {0} {1!r} generated by module {2!r} for ID {3!r}.".format(complaint[
+                    "type"], complaint["message"], checkmodule, complaint["id"]), file=sys.stderr)
+                nr_errors += 1
+
+    # Now check for missing errors and warnings
+    for eid in inputids:
+        # Regex for test_sdsc special ids ("stuff.sdsc.expect.warning.a-an.1")
+        findExpect = re.search(r'.*sdsc\.expect\.([^.]+)\.([^.]+)(?:\..*|$)', eid)
+        if not findExpect:
+            continue
+        found = False
+        if findExpect.group(2) in complaints:
+            found = sum(complaint["id"] == eid and complaint["type"] == findExpect.group(
+                1) for complaint in complaints[findExpect.group(2)])
+
+        if not found:
+            print("Expected {0} for ID {1!r} generated by module {2!r}.".format(
+                findExpect.group(1), eid, findExpect.group(2)), file=sys.stderr)
+            nr_errors += 1
+
+    if nr_errors > 0:
+        pytest.fail(msg="Test {0!r} failed with {1} errors!".format(
+            os.path.basename(xmltestcase), nr_errors), pytrace=False)
